2.11.0b

+++ InitialDLauncher/Form/frmLauncher.vb
@@ -5,7 +5,6 @@ Imports PluginContract
 
 Public Class frmLauncher
 
-    Dim widthx As Integer = 43
     Dim drag As Boolean
     Dim mousex As Integer
     Dim mousey As Integer
@@ -14,7 +13,7 @@ Public Class frmLauncher
     Dim threadU As Thread
     Public shadow As Dropshadow
     Dim curVer As Integer = 39
-    Public buildDate As String = "18/06/2018"
+    Public buildDate As String = "19/06/2018"
 
     Dim id6AppData As String = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "TeknoParrot\SBUU_card.bin")
     Dim id7AppData As String = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "TeknoParrot\SBYD_card.bin")
@@ -32,10 +31,6 @@ Public Class frmLauncher
     Dim selPath As String = String.Empty
     Dim lastGame As Integer = 0
     Public Shared isGameRunning As Boolean = False
-    'Public Shared hideMe As Boolean = False
-    'Public Shared endMe As Boolean = False
-
-    Dim pattern As String = Nothing
 
     Public plugins As ICollection(Of iPlugin) = PluginLoader.LoadPlugins("Plugins")
 
@@ -248,6 +243,13 @@ Public Class frmLauncher
                     isGameRunning = False
                     Exit Sub
                 ElseIf result = DialogResult.Yes Then
+                    If My.Settings.ExtraLaunchOptions.Contains(",") Then
+                        For Each item In My.Settings.ExtraLaunchOptions.Split(",")
+                            Dim psi = New ProcessStartInfo With {.FileName = "CMD", .Arguments = String.Format("/C start """" ""{0}""", item), .WorkingDirectory = Path.GetDirectoryName(item), .UseShellExecute = True, .CreateNoWindow = False, .WindowStyle = ProcessWindowStyle.Normal}
+                            Process.Start(psi)
+                        Next
+                    End If
+
                     selPath = CardPath
                     lastGame = GameID
                     Me.WindowState = FormWindowState.Minimized
@@ -261,6 +263,13 @@ Public Class frmLauncher
                     End If
                 End If
             Else
+                If My.Settings.ExtraLaunchOptions.Contains(",") Then
+                    For Each item In My.Settings.ExtraLaunchOptions.Split(",")
+                        Dim psi = New ProcessStartInfo With {.FileName = "CMD", .Arguments = String.Format("/C start """" ""{0}""", item), .WorkingDirectory = Path.GetDirectoryName(item), .UseShellExecute = True, .CreateNoWindow = False, .WindowStyle = ProcessWindowStyle.Normal}
+                        Process.Start(psi)
+                    Next
+                End If
+
                 selPath = CardPath
                 lastGame = GameID
                 Me.WindowState = FormWindowState.Minimized
@@ -294,7 +303,12 @@ Public Class frmLauncher
                 Dim extension As String = Path.GetExtension(file).Replace(".", "").ToLower
                 Dim destination As String = Path.Combine(My.Application.Info.DirectoryPath, "ID" & CardVersion & "_CARD\" & fileNameOnly)
                 If GetCardVersion(GetHex(fileName, &H50, 2)) = CardVersion Then
-                    IO.File.Move(fileName, destination)
+                    If IO.File.Exists(destination) Then
+                        IO.File.Delete(destination)
+                        IO.File.Move(fileName, destination)
+                    Else
+                        IO.File.Move(fileName, destination)
+                    End If
                 End If
             Next
 
@@ -307,7 +321,12 @@ Public Class frmLauncher
                             Dim extension As String = Path.GetExtension(file).Replace(".", "").ToLower
                             Dim destination As String = Path.Combine(My.Application.Info.DirectoryPath, "ID" & CardVersion & "_CARD\" & fileNameOnly)
                             If GetCardVersion(GetHex(fileName, &H14, 2)) = 6 Then
-                                IO.File.Move(fileName, destination)
+                                If IO.File.Exists(destination) Then
+                                    IO.File.Delete(destination)
+                                    IO.File.Move(fileName, destination)
+                                Else
+                                    IO.File.Move(fileName, destination)
+                                End If
                             End If
                         Next
                     End If
@@ -319,7 +338,12 @@ Public Class frmLauncher
                             Dim extension As String = Path.GetExtension(file).Replace(".", "").ToLower
                             Dim destination As String = Path.Combine(My.Application.Info.DirectoryPath, "ID" & CardVersion & "_CARD\" & fileNameOnly)
                             If GetCardVersion(GetHex(fileName, &H14, 2)) = 7 Then
-                                IO.File.Move(fileName, destination)
+                                If IO.File.Exists(destination) Then
+                                    IO.File.Delete(destination)
+                                    IO.File.Move(fileName, destination)
+                                Else
+                                    IO.File.Move(fileName, destination)
+                                End If
                             End If
                         Next
                     End If
@@ -331,7 +355,12 @@ Public Class frmLauncher
                             Dim extension As String = Path.GetExtension(file).Replace(".", "").ToLower
                             Dim destination As String = Path.Combine(My.Application.Info.DirectoryPath, "ID" & CardVersion & "_CARD\" & fileNameOnly)
                             If GetCardVersion(GetHex(fileName, &H14, 2)) = 8 Then
-                                IO.File.Move(fileName, destination)
+                                If IO.File.Exists(destination) Then
+                                    IO.File.Delete(destination)
+                                    IO.File.Move(fileName, destination)
+                                Else
+                                    IO.File.Move(fileName, destination)
+                                End If
                             End If
                         Next
                     End If
@@ -536,50 +565,6 @@ Public Class frmLauncher
         Me.Cursor = Cursors.Default
         lblLogout.ForeColor = Color.White
     End Sub
-
-    Private Sub frmLauncher_KeyPress(sender As Object, e As KeyPressEventArgs) Handles MyBase.KeyPress
-        pattern = pattern & e.KeyChar
-    End Sub
-
-    'Private Sub Timer2_Tick(sender As Object, e As EventArgs) Handles Timer2.Tick
-    '    Try
-    '        If Not cheat Then
-    '            If pattern.Substring(pattern.Length - 13) = "imnoobcheater" Then
-    '                My.Computer.Audio.Play(My.Resources.play, AudioPlayMode.Background)
-    '                cheat = True
-    '            End If
-    '        End If
-
-    '        If Not debug Then
-    '            If pattern.Substring(pattern.Length - 7) = "debugon" Then
-    '                My.Computer.Audio.Play(My.Resources.play, AudioPlayMode.Background)
-    '                My.Settings.DebugMode = True
-    '                My.Settings.Save()
-    '                debug = True
-    '                lblDebug.Visible = debug
-    '            End If
-    '        End If
-
-    '        If debug Then
-    '            If pattern.Substring(pattern.Length - 8) = "debugoff" Then
-    '                My.Computer.Audio.Play(My.Resources.play, AudioPlayMode.Background)
-    '                My.Settings.DebugMode = False
-    '                My.Settings.Save()
-    '                debug = False
-    '                lblDebug.Visible = debug
-    '            End If
-    '        End If
-
-    '        If hideMe Then
-    '            Me.Hide()
-    '        Else
-    '            Me.Show()
-    '        End If
-
-    '        If endMe Then Me.Close()
-    '    Catch ex As Exception
-    '    End Try
-    'End Sub
 
     Private Sub Timer1_Tick(sender As Object, e As EventArgs) Handles Timer1.Tick
         Try
